+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added `gradient` option to `drawRect()` that specifies the start and end color
 - added `loadProgress()` that returns a `0.0 - 1.0` that indicates current asset loading progress
 - added `kaboom()` option `loadingScreen` where you can turn off the default loading screen
+- added `drawMasked()` and `drawSubtracted()`
 
 ### v2000.2.2